Backend data changes

+++ Entities/Base/Entity.cs
@@ -1,6 +1,6 @@
 namespace fin_app_backend.Entities.Base
 {
-  public abstract class Entity : EntityBase<int>
+  public abstract class Entity : EntityBase<long>
   {
   }
 }
